Add documentation & apply formatting
